+++ src/Domain/Data/Entities/Post.cs
@@ -5,9 +5,9 @@ namespace Mublog.Server.Domain.Data.Entities
 {
     public class Post : BaseEntity
     {
-        public int PublicId { get; set; }
+        public long PublicId { get; set; }
         public string Content { get; set; }
-        public int OwnerId { get; set; }
+        public long OwnerId { get; set; }
         public Profile Owner { get; set; }
         public DateTime PostEditedDate { get; set; }
         public ICollection<Comment> Comments { get; set; }
