+++ src/Infrastructure/Services/Interfaces/IJwtService.cs
@@ -6,7 +6,7 @@ namespace Mublog.Server.Infrastructure.Services.Interfaces
 {
     public interface IJwtService
     {
-        JwtSecurityToken GetToken(string username, int profileId, int accountId, string email = "",  IEnumerable<Claim> additionalClaims = null);
-        string GetTokenString(string username, int profileId, int accountId, string email = "",  IEnumerable<Claim> additionalClaims = null);
+        JwtSecurityToken GetToken(string username, long profileId, long accountId, string email = "",  IEnumerable<Claim> additionalClaims = null);
+        string GetTokenString(string username, long profileId, long accountId, string email = "",  IEnumerable<Claim> additionalClaims = null);
     }
 }
